// src/app/core/api/generated/model/salaryStructureDto.ts
/**
 * Clean Architecture WebAPI
 * An example of Clean Architecture WebAPI with Swagger and JWT Authentication
 *
 * The version of the OpenAPI document: v1
 * Contact: kamrantajerbashi@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface SalaryStructureDto { 
    id?: string;
    employeeId?: string;
    baseSalary?: number;
    hourlyRate?: number;
    overtimeRate?: number;
    effectiveFrom?: string;
    effectiveTo?: string | null;
    isActive?: boolean;
}